reafactor: TXToggleButton -> TXCheckButton 리팩터링 - #243
